End-to-End-Observability-01/app/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
prometheus-client==0.20.0
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-logging==0.45b0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
pydantic==2.7.1